pyproject.toml
[project]
name = "swarm-playground"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "HoangNguyen689", email = "ndhoang.bk.hedspi@gmail.com"},
]
dependencies = [
    "openai>=1.54.4",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "swarm @ git+https://github.com/openai/swarm.git",
    "instructor>=1.6.4",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
